chore: rename refactoring and after test clean up fixes
